:truck: Refactor : re-organize domain directory
