--- src/main/java/com/run_us/server/domains/running/domain/RunningType.java
@@ -1,6 +1,0 @@
-package com.run_us.server.domains.running.domain;
-
-public enum RunningType {
-  SINGLE,
-  GROUP
-}
