+++ src/main/java/com/run_us/server/domains/running/domain/record/PersonalRecord.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.running.domain;
+package com.run_us.server.domains.running.domain.record;
 
 import com.run_us.server.global.common.DateAudit;
 import jakarta.persistence.Column;
@@ -36,14 +36,37 @@ public class PersonalRecord extends DateAudit {
   @Column(name = "record_data", nullable = false)
   private String recordData;
 
-  @Enumerated(EnumType.STRING)
-  private RunningType runningType;
+  @Column(name = "distance")
+  private Integer runningDistanceInMeters;
 
+  @Column(name = "duration")
+  private Integer runningDurationInMilliseconds;
+
+  @Column(name = "avg_pace")
+  private Integer averagePaceInMilliseconds;
+
+  /***
+   * 개인 기록 객체를 위한 생성자
+   * @param runningId 러닝 id
+   * @param userId 유저 id
+   * @param recordData 기록 데이터
+   * @param runningDistanceInMeters 러닝 거리(미터)
+   * @param averagePaceInMilliseconds 평균 페이스(밀리초)
+   * @param runningDurationInMilliseconds 러닝 시간(밀리초)
+   */
   @Builder
-  public PersonalRecord(Long runningId, Long userId, String recordData, RunningType runningType) {
+  public PersonalRecord(
+      Long runningId,
+      Long userId,
+      String recordData,
+      Integer runningDistanceInMeters,
+      Integer runningDurationInMilliseconds,
+      Integer averagePaceInMilliseconds) {
     this.runningId = runningId;
     this.userId = userId;
     this.recordData = recordData;
-    this.runningType = runningType;
+    this.runningDistanceInMeters = runningDistanceInMeters;
+    this.runningDurationInMilliseconds = runningDurationInMilliseconds;
+    this.averagePaceInMilliseconds = averagePaceInMilliseconds;
   }
 }
