+++ src/main/java/com/run_us/server/domains/running/domain/record/RunningType.java
@@ -1,0 +1,6 @@
+package com.run_us.server.domains.running.domain.record;
+
+public enum RunningType {
+  SINGLE,
+  GROUP
+}
